add error handling

+++ src/views/HomeView.vue
@@ -1,5 +1,6 @@
 <script setup>
 import { ref, onMounted } from 'vue'
+import router from "@/router";
 import NiveaubepalingAPI from "@/services/niveaubepalingAPI.js";
 import ItemComponent from "@/components/ItemComponent.vue";
 
@@ -34,7 +35,7 @@ onMounted(() => {
           if (resList.data?.length > 0) {            
             listsData.value = resList.data
             selectedListId.value = resList.data[0].id
-            getListItems(resList.data[0].id)
+            getListItems(selectedListId.value)
           }
           setTimeout(() => {
                 spinner.hide()
@@ -42,7 +43,7 @@ onMounted(() => {
         }
       )
     } 
-  ).catch()
+  ).catch(()=> router.push({ name: 'error' }))
 })
 
 const getListItems = (id) => {
@@ -56,7 +57,7 @@ const getListItems = (id) => {
         }
       )
     } 
-  )
+  ).catch(() => router.push({ name: 'error' }))
 }
 
   const addNewItem = (item) => {
@@ -69,12 +70,9 @@ const getListItems = (id) => {
         res.json().then(
           () => {
             document.getElementById('new-item-input').value = ''
-            getListItems(selectedListId.value)           
+            getListItems(selectedListId.value)      
           }
-        ).catch((err) => {
-          // console.log('i am here', err)
-
-        })
+        ).catch(() => router.push({ name: 'error' }))
       }) 
   }
 
@@ -91,68 +89,44 @@ const getListItems = (id) => {
         })
       }) 
   }
-    function prepareDeleteItem (payload) {
-      itemToBeDeletedId.value = payload
-    }
+  function prepareDeleteItem (payload) {
+    itemToBeDeletedId.value = payload
+  }
 
-    function deleteItem() {
-      console.log('iiiiiiiiiiiiiiiiiiiiiiiiiiiiii' );
+  function deleteItem() {
+    if (itemToBeDeletedId === null) return
     NiveaubepalingAPI.deleteItem( selectedListId.value, itemToBeDeletedId.value)
       .then((res) => {
         res.json().then(
           () => getListItems(selectedListId.value)
-        ).catch((err) => {
-          // console.log('i am here', err)
-        })
+        ).catch(() => router.push({ name: 'error' }))
       }) 
   }
 </script>
 
 <template>
   <main class="container shadow-sm p-5">
-    <svg xmlns="http://www.w3.org/2000/svg" style="display: none;">
-  <symbol id="check-circle-fill" fill="currentColor" viewBox="0 0 16 16">
-    <path d="M16 8A8 8 0 1 1 0 8a8 8 0 0 1 16 0zm-3.97-3.03a.75.75 0 0 0-1.08.022L7.477 9.417 5.384 7.323a.75.75 0 0 0-1.06 1.06L6.97 11.03a.75.75 0 0 0 1.079-.02l3.992-4.99a.75.75 0 0 0-.01-1.05z"/>
-  </symbol>
-  <symbol id="info-fill" fill="currentColor" viewBox="0 0 16 16">
-    <path d="M8 16A8 8 0 1 0 8 0a8 8 0 0 0 0 16zm.93-9.412-1 4.705c-.07.34.029.533.304.533.194 0 .487-.07.686-.246l-.088.416c-.287.346-.92.598-1.465.598-.703 0-1.002-.422-.808-1.319l.738-3.468c.064-.293.006-.399-.287-.47l-.451-.081.082-.381 2.29-.287zM8 5.5a1 1 0 1 1 0-2 1 1 0 0 1 0 2z"/>
-  </symbol>
-  <symbol id="exclamation-triangle-fill" fill="currentColor" viewBox="0 0 16 16">
-    <path d="M8.982 1.566a1.13 1.13 0 0 0-1.96 0L.165 13.233c-.457.778.091 1.767.98 1.767h13.713c.889 0 1.438-.99.98-1.767L8.982 1.566zM8 5c.535 0 .954.462.9.995l-.35 3.507a.552.552 0 0 1-1.1 0L7.1 5.995A.905.905 0 0 1 8 5zm.002 6a1 1 0 1 1 0 2 1 1 0 0 1 0-2z"/>
-  </symbol>
-</svg>
-
- <!-- todo: change the above code place -->
-
-      <div id="spinner-overlay" class="modal" tabindex="-1" aria-labelledby="spinnerModalLabel" aria-hidden="true">
-        <div class="d-flex align-items-center justify-content-center h-100 modal-dialog">
-          <div class="spinner-border text-primary" role="status">
-            <span class="visually-hidden" id="spinnerModalLabel">Loading...</span>
-          </div>
+    <div id="spinner-overlay" class="modal" tabindex="-1" aria-labelledby="spinnerModalLabel" aria-hidden="true">
+      <div class="d-flex align-items-center justify-content-center h-100 modal-dialog">
+        <div class="spinner-border text-primary" role="status">
+          <span class="visually-hidden" id="spinnerModalLabel">Loading...</span>
         </div>
       </div>
-
-      <div id="delete-overlay" class="modal fade" tabindex="-1" aria-labelledby="confirmModalLabel" aria-hidden="true">
-        <div class="modal-dialog">
-          <div class="modal-content">
-            <div class="modal-header">
-              <h5 class="modal-title" id="confirmModalLabel">Confirm</h5>
-            </div>
-            <div class="modal-body">
-              Are you sure you want to delete this item?
-            </div>
-            <div class="modal-footer">
-              <button id="cancel-delete" type="button" class="btn btn-primary" data-bs-dismiss="modal" @click="()=> itemToBeDeletedId = null">No</button>
-              <button id="confirm-delete" type="button" class="btn btn-danger" data-bs-dismiss="modal" @click="deleteItem">Yes</button>
-            </div>
+    </div>
+    <div id="delete-overlay" class="modal fade" tabindex="-1" aria-labelledby="confirmModalLabel" aria-hidden="true">
+      <div class="modal-dialog">
+        <div class="modal-content">
+          <div class="modal-header">
+            <h5 class="modal-title" id="confirmModalLabel">Confirm</h5>
+          </div>
+          <div class="modal-body">
+            Are you sure you want to delete this item?
+          </div>
+          <div class="modal-footer">
+            <button id="cancel-delete" type="button" class="btn btn-primary" data-bs-dismiss="modal" @click="()=> itemToBeDeletedId = null">No</button>
+            <button id="confirm-delete" type="button" class="btn btn-danger" data-bs-dismiss="modal" @click="deleteItem">Yes</button>
           </div>
         </div>
-      </div>
-
-    <div class="alert alert-success d-flex align-items-center" role="alert">
-      <svg class="bi flex-shrink-0 me-2" width="24" height="24" role="img" aria-label="Success:"><use xlink:href="#check-circle-fill"/></svg>
-      <div>
-        a new item created successfully
       </div>
     </div>
     <div class="row justify-content-between g-3">
